12/04/2022 17:41

+++ Main.cpp
@@ -40,7 +40,7 @@ int main()
 
       vsEPC = pBadge->GetListeEPC();
 
-      cout << "NbEPCLues : " << NbEPCLues << endl;
+      cout << "NbEPCLues : " << (int)NbEPCLues << endl;
       
       if( NbEPCLues == 0)
       {
